implement room manager

+++ src/main/java/com/hotel/demo/dto/EconomyRoomReport.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 
 public class EconomyRoomReport extends RoomReport {
 
-    EconomyRoomReport(BigDecimal expectedRevenue, Integer usage) {
+    EconomyRoomReport(BigDecimal expectedRevenue, Long usage) {
         super(expectedRevenue, usage);
     }
 }
