+++ src/main/java/com/hotel/demo/dto/RoomOccupancyReport.java
@@ -11,7 +11,7 @@ public class RoomOccupancyReport {
     private RoomReport premiumRoomData;
     private RoomReport economyRoomData;
 
-    public static RoomOccupancyReport of(Integer premiumUsage, BigDecimal premiumRevenue, Integer economyUsage, BigDecimal economyRevenue) {
+    public static RoomOccupancyReport of(Long premiumUsage, BigDecimal premiumRevenue, Long economyUsage, BigDecimal economyRevenue) {
         return RoomOccupancyReport.builder()
             .economyRoomData(RoomReport.builder().expectedRevenue(economyRevenue).usage(economyUsage).build())
             .premiumRoomData(RoomReport.builder().expectedRevenue(premiumRevenue).usage(premiumUsage).build())
